Change OSX to macOS in helpfiles and builders and repository and code.
This fixes #334

+++ macos/build.sh
@@ -1,78 +1,96 @@
 #!/bin/bash
 
 
-# Synchronize and build libbibledit on macOS for macOS.
+echo Synchronize and build libbibledit on macOS for macOS.
 
 
-# Take the relevant source code for building Bibledit for macOS.
-# Put it in a temporal location.
-# The purpose by doing so is not to have duplicated source code for the bibledit library.
-# This approach does not clutter the bibledit git repository with the built files.
+echo Take the relevant source code for building Bibledit for macOS.
+echo Put it in a temporal location.
+echo Purpose: Not to have duplicated source code for the bibledit library.
+echo This does not clutter the bibledit git repository with the build files.
 MACOSSOURCE=`dirname $0`
 cd $MACOSSOURCE
+if [ $? != 0 ]; then exit; fi
 BIBLEDITMACOS=/tmp/bibledit-macos
 echo Synchronizing relevant source code to $BIBLEDITMACOS
 mkdir -p $BIBLEDITMACOS
 rm $BIBLEDITMACOS/* 2> /dev/null
 rsync --archive --delete ../lib $BIBLEDITMACOS/
+if [ $? != 0 ]; then exit; fi
 rsync --archive --delete ../macos $BIBLEDITMACOS/
+if [ $? != 0 ]; then exit; fi
 
 
-# From now on the working directory is the temporal location.
+echo From now on the working directory is the temporal location.
 cd $BIBLEDITMACOS/macos
+if [ $? != 0 ]; then exit; fi
 
 
 pushd webroot
+if [ $? != 0 ]; then exit; fi
 
 
-# Sychronizes the libbibledit data files in the source tree to macOS and cleans them up.
+echo Sychronize the libbibledit data files in the source tree to macOS.
 rsync -a --delete ../../lib/ .
+if [ $? != 0 ]; then exit; fi
+echo Dist-clean the Bibledit library.
 ./configure
+if [ $? != 0 ]; then exit; fi
 make distclean
+if [ $? != 0 ]; then exit; fi
 rm config.h
 
 
-# Xcode's toolchain for C and C++.
+echo Export the Xcode toolchain for C and C++.
 export CC="/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/clang"
 export CXX="/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/clang++"
 
 
-# Xcode's macOS SDK.
+echo The Xcode macOS SDK.
 SDK=`xcrun --show-sdk-path`
 
 
-# Configure Bibledit in client mode.
+echo Configure Bibledit in client mode.
 ./configure --enable-mac
+if [ $? != 0 ]; then exit; fi
 echo 9876 > config/network-port
 
 
-# Update the Makefile.
-# The -mmmacosx-version-min is for fixing: ld: warning: object file (libbibledit.a(bibledit.o)) was built for newer OSX version (10.11) than being linked (10.10)
+echo Update the Makefile.
+echo The -mmmacosx-version-min is for fixing: ld: warning:
+echo object file libbibledit.a was built for newer OSX version than being linked.
 # sed -i.bak 's#\`xml2-config --cflags\`#-I/usr/include/libxml2#g' Makefile
 sed -i.bak 's#-pedantic#-mmacosx-version-min=10.10\ -isysroot\ '$SDK'#g' Makefile
+if [ $? != 0 ]; then exit; fi
 sed -i.bak 's#/opt/local/include#. -I..#g' Makefile
-# sed -i.bak 's#\`xml2-config --libs\`#-lxml2 -lz -lpthread -liconv -lm#g' Makefile
+if [ $? != 0 ]; then exit; fi
 sed -i.bak 's#/opt/local/lib#.#g' Makefile
-sed -i.bak 's#-L.#-L. -L../lib#g' Makefile
+if [ $? != 0 ]; then exit; fi
 sed -i.bak '/SWORD_CFLAGS =/d' Makefile
+if [ $? != 0 ]; then exit; fi
 sed -i.bak '/SWORD_LIBS =/d' Makefile
+if [ $? != 0 ]; then exit; fi
+sed -i.bak '/OPENSSL_LIBS =/d' Makefile
+if [ $? != 0 ]; then exit; fi
 
 
-# Update the configuration.
+echo Update the configuration.
 sed -i.bak '/CONFIG_ENABLE_FILE_UPLOAD/d' config/config.h
-sed -i.bak '/HAVE_SWORD/d' config.h
+if [ $? != 0 ]; then exit; fi
 rm config/*.bak
 
 
-# Build the Bibledit library.
+echo Build the Bibledit library.
 make -j `sysctl -n hw.logicalcpu_max`
+if [ $? != 0 ]; then exit; fi
 
 
-# Save the header file.
+echo Save the header file.
 cp library/bibledit.h ../macos
+if [ $? != 0 ]; then exit; fi
 
 
-# Clean out stuff no longer needed.
+echo Clean out stuff no longer needed.
 find . -name "*.h" -delete
 find . -name "*.cpp" -delete
 find . -name "*.c" -delete
@@ -101,13 +119,13 @@ rm -rf unittests
 
 
 popd
+if [ $? != 0 ]; then exit; fi
 
 
-
-# Build the app.
+echo Build the app.
 cd $BIBLEDITMACOS/macos
 xcodebuild
-
+if [ $? != 0 ]; then exit; fi
 
 
 echo To graphically build the app for macOS, open the project in Xcode:
